requirements.txt
Django==5.0.2
mercadopago==2.2.1
Pillow==10.2.0
